feat: move exceptions into subfolders
