+++ src/safeds/data/tabular/containers/_table.py
@@ -16,7 +16,7 @@ from scipy import stats
 from safeds.data.image.containers import Image
 from safeds.data.image.typing import ImageFormat
 from safeds.data.tabular.typing import ColumnType, Schema
-from safeds.exceptions import (
+from safeds.data.tabular.exceptions import (
     ColumnLengthMismatchError,
     ColumnSizeError,
     DuplicateColumnNameError,
