+++ src/safeds/data/tabular/exceptions/_exceptions.py
