+++ src/safeds/data/tabular/transformation/_imputer.py
@@ -8,7 +8,7 @@ from sklearn.impute import SimpleImputer as sk_SimpleImputer
 from safeds.data.tabular.containers import Table
 from safeds.data.tabular.transformation._table_transformer import TableTransformer
 from safeds.data.tabular.typing import ImputerStrategy
-from safeds.exceptions import TransformerNotFittedError, UnknownColumnNameError
+from safeds.data.tabular.exceptions import TransformerNotFittedError, UnknownColumnNameError
 
 
 class Imputer(TableTransformer):
