+++ src/safeds/ml/exceptions/__init__.py
@@ -1,0 +1,15 @@
+from ._exceptions import (
+    DatasetContainsTargetError,
+    DatasetMissesFeaturesError,
+    LearningError,
+    ModelNotFittedError,
+    PredictionError,
+)
+
+__all__ = [
+    "DatasetContainsTargetError",
+    "DatasetMissesFeaturesError",
+    "LearningError",
+    "ModelNotFittedError",
+    "PredictionError",
+]
