+++ tests/safeds/data/tabular/containers/_column/test_boxplot.py
@@ -3,7 +3,7 @@ import matplotlib.pyplot as plt
 import pandas as pd
 import pytest
 from safeds.data.tabular.containers import Table
-from safeds.exceptions import NonNumericColumnError
+from safeds.data.tabular.exceptions import NonNumericColumnError
 
 
 def test_boxplot_complex() -> None:
