+++ tests/safeds/data/tabular/containers/_column/test_idness.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytest
 from safeds.data.tabular.containers import Column
-from safeds.exceptions import ColumnSizeError
+from safeds.data.tabular.exceptions import ColumnSizeError
 
 
 @pytest.mark.parametrize(
