+++ tests/safeds/data/tabular/containers/_column/test_minimum.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytest
 from safeds.data.tabular.containers import Table
-from safeds.exceptions import NonNumericColumnError
+from safeds.data.tabular.exceptions import NonNumericColumnError
 
 
 def test_minimum_invalid() -> None:
