+++ tests/safeds/data/tabular/containers/_table/test_add_row.py
@@ -2,7 +2,7 @@ import pandas as pd
 from _pytest.python_api import raises
 from safeds.data.tabular.containers import Row, Table
 from safeds.data.tabular.typing import Integer, Schema, String
-from safeds.exceptions import SchemaMismatchError
+from safeds.data.tabular.exceptions import SchemaMismatchError
 
 
 def test_add_row_valid() -> None:
