+++ tests/safeds/data/tabular/containers/_table/test_from_columns.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytest
 from safeds.data.tabular.containers import Column, Table
-from safeds.exceptions import MissingDataError
+from safeds.data.tabular.exceptions import MissingDataError
 
 from tests.helpers import resolve_resource_path
