+++ tests/safeds/data/tabular/containers/_table/test_from_rows.py
@@ -1,6 +1,6 @@
 import pytest
 from safeds.data.tabular.containers import Row, Table
-from safeds.exceptions import MissingDataError
+from safeds.data.tabular.exceptions import MissingDataError
 
 from tests.helpers import resolve_resource_path
