+++ tests/safeds/data/tabular/containers/_table/test_keep_only_columns.py
@@ -1,6 +1,6 @@
 import pytest
 from safeds.data.tabular.containers import Table
-from safeds.exceptions import UnknownColumnNameError
+from safeds.data.tabular.exceptions import UnknownColumnNameError
 
 from tests.helpers import resolve_resource_path
