+++ tests/safeds/data/tabular/containers/_table/test_table_add_column.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytest
 from safeds.data.tabular.containers import Column, Table
-from safeds.exceptions import ColumnSizeError, DuplicateColumnNameError
+from safeds.data.tabular.exceptions import ColumnSizeError, DuplicateColumnNameError
 
 from tests.helpers import resolve_resource_path
