+++ tests/safeds/data/tabular/containers/test_row.py
@@ -3,7 +3,7 @@ from typing import Any
 import pytest
 from safeds.data.tabular.containers import Row, Table
 from safeds.data.tabular.typing import ColumnType, Integer, Schema, String
-from safeds.exceptions import UnknownColumnNameError
+from safeds.data.tabular.exceptions import UnknownColumnNameError
 
 
 class TestInit:
