+++ tests/safeds/data/tabular/containers/test_tagged_table.py
@@ -1,6 +1,6 @@
 import pytest
 from safeds.data.tabular.containers import Column, Table, TaggedTable
-from safeds.exceptions import UnknownColumnNameError
+from safeds.data.tabular.exceptions import UnknownColumnNameError
 
 
 @pytest.fixture()
