+++ tests/safeds/data/tabular/transformation/test_imputer.py
@@ -2,7 +2,7 @@ import pytest
 from safeds.data.tabular.containers import Column, Table
 from safeds.data.tabular.transformation import Imputer
 from safeds.data.tabular.typing import ImputerStrategy
-from safeds.exceptions import TransformerNotFittedError, UnknownColumnNameError
+from safeds.data.tabular.exceptions import TransformerNotFittedError, UnknownColumnNameError
 
 
 class TestFit:
