+++ tests/safeds/data/tabular/transformation/test_one_hot_encoder.py
@@ -1,7 +1,7 @@
 import pytest
 from safeds.data.tabular.containers import Column, Table
 from safeds.data.tabular.transformation import OneHotEncoder
-from safeds.exceptions import TransformerNotFittedError, UnknownColumnNameError
+from safeds.data.tabular.exceptions import TransformerNotFittedError, UnknownColumnNameError
 
 
 class TestFit:
