+++ tests/safeds/ml/regression/test_regressor.py
@@ -5,8 +5,8 @@ from typing import TYPE_CHECKING
 import pandas as pd
 import pytest
 from safeds.data.tabular.containers import Column, Table, TaggedTable
-from safeds.exceptions import (
-    ColumnLengthMismatchError,
+from safeds.data.tabular.exceptions import ColumnLengthMismatchError
+from safeds.ml.exceptions import (
     DatasetContainsTargetError,
     DatasetMissesFeaturesError,
     LearningError,
